Remove struct ProtoArrayForkChoice
