+++ consensus/fork_choice/Cargo.toml
@@ -11,6 +11,7 @@ ethereum_ssz_derive = { workspace = true }
 logging = { workspace = true }
 metrics = { workspace = true }
 proto_array = { workspace = true }
+serde = { workspace = true }
 state_processing = { workspace = true }
 superstruct = { workspace = true }
 tracing = { workspace = true }
