+++ consensus/proto_array/src/error.rs
@@ -52,7 +52,6 @@ pub enum Error {
         block_root: Hash256,
         parent_root: Hash256,
     },
-    InvalidEpochOffset(u64),
     Arith(ArithError),
 }
