+++ consensus/proto_array/src/justified_balances.rs
@@ -1,7 +1,8 @@
 use safe_arith::{ArithError, SafeArith};
+use serde::{Deserialize, Serialize};
 use types::{BeaconState, EthSpec};
 
-#[derive(Debug, PartialEq, Clone, Default)]
+#[derive(Debug, PartialEq, Serialize, Deserialize, Clone, Default)]
 pub struct JustifiedBalances {
     /// The effective balances for every validator in a given justified state.
     ///
