+++ consensus/proto_array/src/lib.rs
@@ -1,20 +1,19 @@
+mod compute_deltas;
 mod error;
+mod execution_status;
 pub mod fork_choice_test_definition;
 mod justified_balances;
 mod proto_array;
-mod proto_array_fork_choice;
 mod ssz_container;
 
+pub use crate::execution_status::ExecutionStatus;
 pub use crate::justified_balances::JustifiedBalances;
-pub use crate::proto_array::{InvalidationOperation, calculate_committee_fraction};
-pub use crate::proto_array_fork_choice::{
-    Block, DisallowedReOrgOffsets, DoNotReOrg, ExecutionStatus, ProposerHeadError,
-    ProposerHeadInfo, ProtoArrayForkChoice, ReOrgThreshold,
+pub use crate::proto_array::{
+    InvalidationOperation, ProtoArray, ProtoNode, calculate_committee_fraction,
 };
 pub use error::Error;
 
 pub mod core {
-    pub use super::proto_array::{ProposerBoost, ProtoArray, ProtoNode};
-    pub use super::proto_array_fork_choice::VoteTracker;
+    pub use super::proto_array::{ProposerBoost, ProtoArray, ProtoNode, VoteTracker};
     pub use super::ssz_container::{SszContainer, SszContainerV17, SszContainerV28};
 }
